requirements.txt
absl-py==0.7.0
alabaster==0.7.12
anaconda-client==1.7.2
anaconda-navigator==1.9.6
anaconda-project==0.8.2
appdirs==1.4.4
asn1crypto==0.24.0
astor==0.7.1
astroid==2.1.0
astropy==3.1.2
atomicwrites==1.3.0
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bkcharts==0.2
blaze==0.11.3
bleach==3.1.0
bokeh==1.0.4
boto==2.49.0
bottle==0.12.18
Bottleneck==1.2.1
certifi==2018.11.29
cffi==1.12.1
chardet==3.0.4
cheroot==8.3.0
CherryPy==18.6.0
Click==7.0
Clockwork==1.2.0
cloudpickle==0.8.0
clyent==1.2.2
cmake==3.17.3
colorama==0.4.1
comtypes==1.1.7
conda==4.6.7
conda-build==3.17.8
conda-verify==3.1.1
contextlib2==0.5.5
cryptography==2.5
cycler==0.10.0
Cython==0.29.5
cytoolz==0.9.0.1
dask==1.1.3
datashape==0.5.4
decorator==4.3.2
defusedxml==0.5.0
distlib==0.3.0
distributed==1.26.0
docopt==0.6.2
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
face-recognition-models==0.3.0
fastcache==1.0.2
filelock==3.0.10
Flask==1.0.2
Flask-Cors==3.0.7
future==0.17.1
futures==3.1.1
gast==0.2.2
gevent==1.4.0
glob2==0.6
goslate==1.5.1
graphviz==0.13.2
greenlet==0.4.15
grpcio==1.19.0
h5py==2.8.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==1.6.1
ipykernel==5.1.0
ipython==7.3.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.8
itsdangerous==1.1.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.text==3.2.0
jdcal==1.4
jedi==0.13.3
Jinja2==2.10
joblib==0.14.1
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.3
jupyterlab-server==0.2.0
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
keyring==18.0.0
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
libarchive-c==2.8
llvmlite==0.27.0
locket==0.2.0
lxml==4.3.1
Markdown==3.0.1
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
menuinst==1.4.14
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
mongoengine==0.20.0
more-itertools==5.0.0
mpmath==1.1.0
msgpack==0.6.1
multipledispatch==0.6.0
mysql-connector-python==8.0.20
mysqlclient==1.4.2.post1
navigator-updater==0.2.1
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
nltk==3.4
nose==1.3.7
notebook==5.7.4
numba==0.42.0
numexpr==2.6.9
numpy==1.16.2
numpydoc==0.8.0
odo==0.5.1
olefile==0.46
openpyxl==2.6.0
packaging==19.0
pandas==0.24.1
pandocfilters==1.4.2
parso==0.3.4
partd==0.3.9
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.1
pbr==5.1.3
pdf2image==1.8.0
pep8==1.7.1
pickleshare==0.7.5
Pillow==5.4.1
pipreqs==0.4.10
pkginfo==1.5.0.1
pluggy==0.9.0
ply==3.11
portend==2.6
praw==6.5.1
prawcore==1.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.0
psutil==5.5.0
py==1.8.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
PyDictionary==1.5.2
pydotplus==2.0.2
pyflakes==2.1.0
Pygments==2.3.1
pylint==2.2.2
pymongo==3.10.1
pyodbc==4.0.26
pyOpenSSL==19.0.0
pyparsing==2.3.1
PySocks==1.6.8
pytesseract==0.3.4
pytest==4.3.0
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.2.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.1
pywin32==223
pywinpty==0.5.5
PyYAML==3.13
pyzmq==18.0.0
QtAwesome==0.5.6
qtconsole==4.4.3
QtPy==1.6.0
requests==2.21.0
rope==0.12.0
ruamel-yaml==0.15.46
scikit-image==0.14.1
scikit-learn==0.23.1
scipy==1.2.1
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.12.0
slimit==0.8.1
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
soupsieve==1.7.1
Sphinx==1.8.4
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.2.18
statsmodels==0.9.0
sympy==1.3
tables==3.4.4
tblib==1.3.2
tempora==3.0.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
Theano==1.0.4
threadpoolctl==2.0.0
toolz==0.9.0
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
unicodecsv==0.14.1
update-checker==0.16
urllib2-file==0.2.1
urllib3==1.24.1
virtualenv==20.0.23
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win-inet-pton==1.1.0
win-unicode-console==0.5
wincertstore==0.2
wrapt==1.11.1
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.3
xlwt==1.3.0
yarg==0.1.9
zc.lockfile==2.0
zict==0.1.3
zipp==3.1.0
